Split up store into files and document it better
